Mireya/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0

python-dotenv==1.1.1
aiohttp>=3.9.0
pydantic==2.11.7
typing_extensions==4.15.0

supabase==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
matplotlib==3.10.6